# openapi.yaml
...
openapi: 3.1.0
info:
  title: Inference Gateway API
  description: |
    The API for interacting with various language models and other AI services.
    OpenAI, Groq, Ollama, and other providers are supported.
    OpenAI compatible API for using with existing clients.
    Unified API for all providers.
  contact:
    name: Inference Gateway
    url: https://inference-gateway.github.io/docs/
  version: 1.0.0
  license:
    name: MIT
    url: https://github.com/thirdwake/inference-gateway/blob/main/LICENSE
servers:
  - url: http://localhost:8080
    description: Default server without version prefix for healthcheck and proxy and points
    x-server-tags: ["Health", "Proxy"]
  - url: http://localhost:8080/v1
    description: Default server with version prefix for listing models and chat completions
    x-server-tags: ["Models", "Completions"]
  - url: https://api.inference-gateway.local/v1
    description: Local server with version prefix for listing models and chat completions
    x-server-tags: ["Models", "Completions"]
...
paths:
  /models:
    get:
      operationId: listModels
      tags:
        - Models
      description: |
        Lists the currently available models, and provides basic information
        about each one such as the owner and availability.
      summary:
        Lists the currently available models, and provides basic information
        about each one such as the owner and availability.
      security:
        - bearerAuth: []
      parameters:
        - name: provider
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Provider"
          description: Specific provider to query (optional)
      responses:
        "200":
          description: List of available models
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListModelsResponse"
              examples:
                allProviders:
                  summary: Models from all providers
                  value:
                    object: "list"
                    data:
                      - id: "openai/gpt-4o"
                        object: "model"
                        created: 1686935002
                        owned_by: "openai"
                        served_by: "openai"
                      - id: "openai/llama-3.3-70b-versatile"
                        object: "model"
                        created: 1723651281
                        owned_by: "groq"
                        served_by: "groq"
                      - id: "cohere/claude-3-opus-20240229"
                        object: "model"
                        created: 1708905600
                        owned_by: "anthropic"
                        served_by: "anthropic"
                      - id: "cohere/command-r"
                        object: "model"
                        created: 1707868800
                        owned_by: "cohere"
                        served_by: "cohere"
                      - id: "ollama/phi3:3.8b"
                        object: "model"
                        created: 1718441600
                        owned_by: "ollama"
                        served_by: "ollama"
                singleProvider:
                  summary: Models from a specific provider
                  value:
                    object: "list"
                    data:
                      - id: "openai/gpt-4o"
                        object: "model"
                        created: 1686935002
                        owned_by: "openai"
                        served_by: "openai"
                      - id: "openai/gpt-4-turbo"
                        object: "model"
                        created: 1687882410
                        owned_by: "openai"
                        served_by: "openai"
                      - id: "openai/gpt-3.5-turbo"
                        object: "model"
                        created: 1677649963
                        owned_by: "openai"
                        served_by: "openai"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
...
  responses:
...
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Authentication is optional by default.
        To enable authentication, set ENABLE_AUTH to true.
        When enabled, requests must include a valid JWT token in the Authorization header.
  schemas:
    Provider:
      type: string
      enum:
        - ollama
        - groq
        - openai
        - cloudflare
        - cohere
        - anthropic
        - deepseek
      x-provider-configs:
        ollama:
          id: "ollama"
          url: "http://ollama:8080/v1"
          auth_type: "none"
          endpoints:
            models:
              name: "list_models"
              method: "GET"
              endpoint: "/models"
            chat:
              name: "chat_completions"
              method: "POST"
              endpoint: "/chat/completions"
        anthropic:
          id: "anthropic"
          url: "https://api.anthropic.com/v1"
          auth_type: "bearer"
          endpoints:
            models:
              name: "list_models"
              method: "GET"
              endpoint: "/models"
            chat:
              name: "chat_completions"
              method: "POST"
              endpoint: "/chat/completions"
        cohere:
          id: "cohere"
          url: "https://api.cohere.ai"
          auth_type: "bearer"
          endpoints:
            models:
              name: "list_models"
              method: "GET"
              endpoint: "/v1/models"
            chat:
              name: "chat_completions"
              method: "POST"
              endpoint: "/compatibility/v1/chat/completions"
        groq:
          id: "groq"
          url: "https://api.groq.com/openai/v1"
          auth_type: "bearer"
          endpoints:
            models:
              name: "list_models"
              method: "GET"
              endpoint: "/models"
            chat:
              name: "chat_completions"
              method: "POST"
              endpoint: "/chat/completions"
        openai:
          id: "openai"
          url: "https://api.openai.com/v1"
          auth_type: "bearer"
          endpoints:
            models:
              name: "list_models"
              method: "GET"
              endpoint: "/models"
            chat:
              name: "chat_completions"
              method: "POST"
              endpoint: "/chat/completions"
        cloudflare:
          id: "cloudflare"
          url: "https://api.cloudflare.com/client/v4/accounts/{ACCOUNT_ID}/ai"
          auth_type: "bearer"
          endpoints:
            models:
              name: "list_models"
              method: "GET"
              endpoint: "/finetunes/public?limit=1000"
            chat:
              name: "chat_completions"
              method: "POST"
              endpoint: "/v1/chat/completions"
        deepseek:
          id: "deepseek"
          url: "https://api.deepseek.com"
          auth_type: "bearer"
          endpoints:
            models:
              name: "list_models"
              method: "GET"
              endpoint: "/models"
            chat:
              name: "chat_completions"
              method: "POST"
              endpoint: "/chat/completions"
...
    Error:
      type: object
      properties:
        error:
          type: string
...
    Model:
      type: object
      description: Common model information
      properties:
        id:
          type: string
        object:
          type: string
        created:
          type: integer
          format: int64
        owned_by:
          type: string
        served_by:
          $ref: "#/components/schemas/Provider"
      required:
        - id
        - object
        - created
        - owned_by
        - served_by
    ListModelsResponse:
      type: object
      description: Response structure for listing models
      properties:
        provider:
          $ref: "#/components/schemas/Provider"
        object:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/Model"
          default: []
      required:
        - object
        - data
...